requirements.txt
snowflake-connector-python
snowflake-snowpark-python
pandas
pyarrow
click